+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Patches "$HOME" for setting ~/.gitconfig in tests
 
 ### Removed
+- Remove: removes `PyaudEnvironmentError` for `EnvironmentError
 - Removes json from `Environ.__repr__`
 
 [1.3.0](https://github.com/jshwi/pyaud/releases/tag/v1.3.0) - 2021-03-18